+++ features/org.jkiss.dbeaver.db.feature/feature.xml
@@ -20,7 +20,7 @@
    <plugin id="org.jkiss.dbeaver.ext.db2.zos" version="0.0.0" />
    <plugin id="org.jkiss.dbeaver.ext.db2.i" version="0.0.0" />
    <plugin id="org.jkiss.dbeaver.ext.databend" version="0.0.0" />
-   <plugin id="org.jkiss.dbeaver.ext.lake" version="0.0.0" />
+   <plugin id="org.jkiss.dbeaver.ext.tidblake" version="0.0.0" />
    <plugin id="org.jkiss.dbeaver.ext.derby" version="0.0.0" />
    <plugin id="org.jkiss.dbeaver.ext.dameng" version="0.0.0"  />
    <plugin id="org.jkiss.dbeaver.ext.duckdb" version="0.0.0" />
