+++ features/org.jkiss.dbeaver.db.ui.feature/feature.xml
@@ -22,7 +22,7 @@
    <plugin id="org.jkiss.dbeaver.ext.tidb.ui" version="0.0.0" />
    <plugin id="org.jkiss.dbeaver.ext.postgresql.ui" version="0.0.0" />
    <plugin id="org.jkiss.dbeaver.ext.databend.ui" version="0.0.0" />
-   <plugin id="org.jkiss.dbeaver.ext.lake.ui" version="0.0.0" />
+   <plugin id="org.jkiss.dbeaver.ext.tidblake.ui" version="0.0.0" />
    <plugin id="org.jkiss.dbeaver.ext.generic.ui" version="0.0.0" />
    <plugin id="org.jkiss.dbeaver.ext.vertica.ui" version="0.0.0" />
    <plugin id="org.jkiss.dbeaver.ext.athena.ui" version="0.0.0" />
