--- plugins/org.jkiss.dbeaver.ext.lake.ui/META-INF/MANIFEST.MF
@@ -1,15 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: DBeaver Lake UI
-Bundle-SymbolicName: org.jkiss.dbeaver.ext.lake.ui;singleton:=true
-Bundle-Version: 1.0.0.qualifier
-Bundle-Release-Date: 20260629
-Require-Bundle: org.jkiss.dbeaver.model,
- org.jkiss.dbeaver.ext.lake,
- org.jkiss.dbeaver.ui,
- org.jkiss.dbeaver.ui.editors.connection,
- org.jkiss.dbeaver.ext.generic.ui
-Bundle-ActivationPolicy: lazy
-Bundle-RequiredExecutionEnvironment: JavaSE-21
-Bundle-Vendor: DBeaver Corp
-Automatic-Module-Name: org.jkiss.dbeaver.ext.lake.ui
