--- plugins/org.jkiss.dbeaver.ext.lake.ui/pom.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>org.jkiss.dbeaver</groupId>
-    <artifactId>plugins</artifactId>
-    <version>1.0.0-SNAPSHOT</version>
-     <relativePath>../</relativePath>
-  </parent>
-  <artifactId>org.jkiss.dbeaver.ext.lake.ui</artifactId>
-  <version>1.0.0-SNAPSHOT</version>
-  <packaging>eclipse-plugin</packaging>
-</project>
