--- plugins/org.jkiss.dbeaver.ext.lake.ui/src/org/jkiss/dbeaver/ext/lake/ui/LakeSSLConfigurator.java
@@ -1,29 +1,0 @@
-/*
- * DBeaver - Universal Database Manager
- * Copyright (C) 2010-2026 DBeaver Corp and others
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jkiss.dbeaver.ext.lake.ui;
-
-import org.eclipse.swt.widgets.Composite;
-import org.jkiss.code.NotNull;
-import org.jkiss.dbeaver.ui.dialogs.net.SSLConfiguratorTrustStoreUI;
-
-public class LakeSSLConfigurator extends SSLConfiguratorTrustStoreUI {
-    @Override
-    public void createControl(@NotNull Composite parent, Object object, @NotNull Runnable propertyChangeListener) {
-        createSSLConfigHint(parent, true, 0);
-        createTrustStoreConfigGroup(parent);
-    }
-}
