--- plugins/org.jkiss.dbeaver.ext.lake/META-INF/MANIFEST.MF
@@ -1,15 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: DBeaver Lake Model
-Bundle-SymbolicName: org.jkiss.dbeaver.ext.lake;singleton:=true
-Bundle-Version: 1.0.0.qualifier
-Bundle-Release-Date: 20260629
-Require-Bundle: org.jkiss.dbeaver.model,
- org.jkiss.dbeaver.ext.generic;visibility:=reexport
-Export-Package: org.jkiss.dbeaver.ext.lake,
- org.jkiss.dbeaver.ext.lake.model
-Import-Package: org.slf4j
-Bundle-ActivationPolicy: lazy
-Bundle-RequiredExecutionEnvironment: JavaSE-21
-Bundle-Vendor: DBeaver Corp
-Automatic-Module-Name: org.jkiss.dbeaver.ext.lake
