--- plugins/org.jkiss.dbeaver.ext.lake/plugin.xml
@@ -1,76 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<?eclipse version="3.2"?>
-
-<plugin>
-
-    <extension point="org.jkiss.dbeaver.generic.meta">
-        <meta id="lake"
-              class="org.jkiss.dbeaver.ext.lake.model.LakeMetaModel"
-              driverClass="com.tidbcloud.jdbc.LakeDriver"
-              dialect="lake"/>
-    </extension>
-
-    <extension point="org.jkiss.dbeaver.dataSourceProvider">
-        <datasource
-                class="org.jkiss.dbeaver.ext.lake.LakeDataSourceProvider"
-                description="Lake"
-                id="lake"
-                parent="generic"
-                label="Lake"
-                icon="icons/lake_icon.png"
-                dialect="lake">
-            <drivers managable="true">
-                <driver
-                    id="lake"
-                    label="Lake"
-                    class="com.tidbcloud.jdbc.LakeDriver"
-                    sampleURL="jdbc:lake://{host}:{port}[/{database}]"
-                    defaultPort="443"
-                    defaultDatabase="default"
-                    icon="icons/lake_icon.png"
-                    iconBig="icons/lake_icon_big.png"
-                    description="TiDB Cloud Lake driver"
-                    supportedConfigurationTypes="MANUAL,URL"
-                    webURL="https://github.com/tidbcloud/lake-jdbc"
-                    categories="sql,analytic,columns">
-                    <file type="jar" path="maven:/com.tidbcloud:lake-jdbc:RELEASE[0.4.6]" load-optional-dependencies="true" bundle="!drivers.lake" />
-                    <file type="license" path="licenses/external/asl-2.0.txt"/>
-                    <file type="jar" path="drivers/lake" bundle="drivers.lake"/>
-
-                    <parameter name="supports-indexes" value="false"/>
-                    <parameter name="supports-references" value="false"/>
-                    <parameter name="supports-truncate" value="false"/>
-                    <parameter name="alter-table-add-column" value="true"/>
-                    <parameter name="supports-multi-insert" value="true"/>
-                    <parameter name="omit-catalog" value="false"/>
-                    <parameter name="omit-schema" value="false"/>
-                    <parameter name="query-get-active-db" value="SELECT CURRENT_DATABASE()"/>
-                    <parameter name="query-set-active-db" value="USE &quot;?&quot;"/>
-                </driver>
-            </drivers>
-
-        </datasource>
-    </extension>
-
-    <extension point="org.jkiss.dbeaver.sqlDialect">
-        <dialect id="lake" parent="generic" class="org.jkiss.dbeaver.ext.lake.model.LakeSQLDialect" label="Lake" description="Lake SQL dialect." icon="icons/lake_icon.png">
-        </dialect>
-    </extension>
-
-    <extension point="org.jkiss.dbeaver.networkHandler">
-        <handler
-                type="config"
-                id="lake-ssl"
-                codeName="SSL"
-                prefix="ssl"
-                label="SSL"
-                description="SSL settings"
-                secured="false"
-                desktop="true"
-                order="2"
-                handlerClass="org.jkiss.dbeaver.model.impl.net.SSLHandlerImpl">
-            <objectType name="org.jkiss.dbeaver.ext.lake.LakeDataSourceProvider"/>
-        </handler>
-    </extension>
-
-</plugin>
