--- plugins/org.jkiss.dbeaver.ext.lake/src/org/jkiss/dbeaver/ext/lake/model/LakeMetaModel.java
@@ -1,132 +1,0 @@
-/*
- * DBeaver - Universal Database Manager
- * Copyright (C) 2010-2026 DBeaver Corp and others
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jkiss.dbeaver.ext.lake.model;
-
-import org.jkiss.code.NotNull;
-import org.jkiss.code.Nullable;
-import org.jkiss.dbeaver.DBDatabaseException;
-import org.jkiss.dbeaver.DBException;
-import org.jkiss.dbeaver.ext.generic.model.GenericCatalog;
-import org.jkiss.dbeaver.ext.generic.model.GenericSchema;
-import org.jkiss.dbeaver.ext.generic.model.GenericTableBase;
-import org.jkiss.dbeaver.ext.generic.model.GenericView;
-import org.jkiss.dbeaver.ext.generic.model.meta.GenericMetaModel;
-import org.jkiss.dbeaver.model.DBPErrorAssistant;
-import org.jkiss.dbeaver.model.DBPEvaluationContext;
-import org.jkiss.dbeaver.model.DBUtils;
-import org.jkiss.dbeaver.model.exec.DBCQueryTransformProvider;
-import org.jkiss.dbeaver.model.exec.DBCQueryTransformType;
-import org.jkiss.dbeaver.model.exec.DBCQueryTransformer;
-import org.jkiss.dbeaver.model.exec.jdbc.JDBCPreparedStatement;
-import org.jkiss.dbeaver.model.exec.jdbc.JDBCResultSet;
-import org.jkiss.dbeaver.model.exec.jdbc.JDBCSession;
-import org.jkiss.dbeaver.model.impl.sql.QueryTransformerLimit;
-import org.jkiss.dbeaver.model.runtime.DBRProgressMonitor;
-import org.jkiss.utils.CommonUtils;
-
-import java.sql.SQLException;
-import java.util.Map;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
-/**
- * Lake meta model
- */
-public class LakeMetaModel extends GenericMetaModel implements DBCQueryTransformProvider {
-
-    private Pattern ERROR_POSITION_PATTERN = Pattern.compile(" SQL([0-9]+)\\:([0-9]+)");
-    private static final String TABLE_DDL = "SHOW CREATE TABLE ";
-
-    public LakeMetaModel() {
-    }
-
-    @Nullable
-    @Override
-    public DBCQueryTransformer createQueryTransformer(@NotNull DBCQueryTransformType type) {
-        if (type == DBCQueryTransformType.RESULT_SET_LIMIT) {
-            return new QueryTransformerLimit(false);
-        }
-        return null;
-    }
-
-    @Override
-    public String getTableDDL(@NotNull DBRProgressMonitor monitor, @NotNull GenericTableBase sourceObject, @NotNull Map<String, Object> options) throws DBException {
-        GenericSchema schema =  sourceObject.getSchema();
-        GenericCatalog catalog =  sourceObject.getCatalog();
-        if (
-            (schema != null && schema.getName().equals("system"))
-            || (catalog != null && catalog.getName().equals("system"))
-        ) {
-            return super.getTableDDL(monitor, sourceObject, options);
-        }
-        try (JDBCSession session = DBUtils.openMetaSession(monitor, sourceObject, "Read Lake object DDL")) {
-            try (JDBCPreparedStatement dbStat = session.prepareStatement(
-                    TABLE_DDL + " " + sourceObject.getFullyQualifiedName(DBPEvaluationContext.DDL))) {
-                try (JDBCResultSet dbResult = dbStat.executeQuery()) {
-                    StringBuilder sql = new StringBuilder();
-                    while (dbResult.nextRow()) {
-                        sql.append(dbResult.getString(2)).append("\n");
-                    }
-                    return sql.toString();
-                }
-            }
-        } catch (SQLException e) {
-            throw new DBDatabaseException(e, sourceObject.getDataSource());
-        }
-    }
-
-    @Override
-    public boolean supportsTableDDLSplit(@NotNull GenericTableBase sourceObject) {
-        return false;
-    }
-
-    @Override
-    public String getViewDDL(@NotNull DBRProgressMonitor monitor, @NotNull GenericView sourceObject, @NotNull Map<String, Object> options) throws DBException {
-        return getTableDDL(monitor, sourceObject, options);
-    }
-
-    @Nullable
-    @Override
-    public DBPErrorAssistant.ErrorPosition getErrorPosition(@NotNull Throwable error) {
-        String message = error.getMessage();
-        if (!CommonUtils.isEmpty(message)) {
-            Matcher matcher = ERROR_POSITION_PATTERN.matcher(message);
-            if (matcher.find()) {
-                DBPErrorAssistant.ErrorPosition pos = new DBPErrorAssistant.ErrorPosition();
-                pos.line = Integer.parseInt(matcher.group(1)) - 1;
-                pos.position = Integer.parseInt(matcher.group(2)) - 1;
-                return pos;
-            }
-        }
-        return null;
-    }
-
-    @Override
-    public boolean isSchemasOptional() {
-        return false;
-    }
-
-    @Override
-    public boolean hasProcedureSupport() {
-        return false;  // driver does not return correct information
-    }
-
-    @Override
-    public boolean hasFunctionSupport() {
-        return false;  // driver does not return correct information
-    }
-}
