+++ plugins/org.jkiss.dbeaver.ext.tidblake.ui/build.properties
