+++ plugins/org.jkiss.dbeaver.ext.tidblake.ui/src/org/jkiss/dbeaver/ext/tidblake/ui/internal/TiDBLakeMessages_zh.properties
