+++ plugins/org.jkiss.dbeaver.ext.tidblake/src/org/jkiss/dbeaver/ext/tidblake/model/TiDBLakeConstants.java
@@ -1,0 +1,22 @@
+/*
+ * DBeaver - Universal Database Manager
+ * Copyright (C) 2010-2026 DBeaver Corp and others
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.jkiss.dbeaver.ext.tidblake.model;
+
+public class TiDBLakeConstants {
+    public static final String JDBC_URL_PREFIX = "jdbc:lake://";
+    public static final String SSL_PARAM = "sslmode"; //$NON-NLS-1$
+}
