+++ plugins/org.jkiss.dbeaver.ext.tidblake/src/org/jkiss/dbeaver/ext/tidblake/model/TiDBLakeDataSource.java
@@ -1,0 +1,70 @@
+/*
+ * DBeaver - Universal Database Manager
+ * Copyright (C) 2010-2026 DBeaver Corp and others
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.jkiss.dbeaver.ext.tidblake.model;
+
+import org.jkiss.code.NotNull;
+import org.jkiss.dbeaver.DBException;
+import org.jkiss.dbeaver.ext.generic.model.GenericDataSource;
+import org.jkiss.dbeaver.model.DBPDataSourceContainer;
+import org.jkiss.dbeaver.model.connection.DBPConnectionConfiguration;
+import org.jkiss.dbeaver.model.exec.DBCException;
+import org.jkiss.dbeaver.model.impl.jdbc.JDBCExecutionContext;
+import org.jkiss.dbeaver.model.net.DBWHandlerConfiguration;
+import org.jkiss.dbeaver.model.runtime.DBRProgressMonitor;
+
+import java.util.Properties;
+
+/**
+ * TiDBLake datasource
+ */
+public class TiDBLakeDataSource extends GenericDataSource {
+
+    public TiDBLakeDataSource(DBRProgressMonitor monitor, DBPDataSourceContainer container, TiDBLakeMetaModel metaModel)
+            throws DBException {
+        super(monitor, container, metaModel, new TiDBLakeSQLDialect());
+    }
+
+
+    @Override
+    protected boolean isPopulateClientAppName() {
+        return false;
+    }
+
+    @NotNull
+    @Override
+    protected Properties getAllConnectionProperties(@NotNull DBRProgressMonitor monitor, @NotNull JDBCExecutionContext context, @NotNull
+    String purpose, @NotNull DBPConnectionConfiguration connectionInfo)
+            throws DBCException {
+        Properties properties = super.getAllConnectionProperties(monitor, context, purpose, connectionInfo);
+
+        final DBWHandlerConfiguration sslConfig = getContainer().getActualConnectionConfiguration().getHandler("tidblake-ssl");
+
+        if (sslConfig != null && sslConfig.isEnabled()) {
+            try {
+                initSSL(monitor, properties);
+            } catch (Exception e) {
+                throw new DBCException("Error configuring SSL certificates", e);
+            }
+        }
+        return properties;
+    }
+
+    private void initSSL(DBRProgressMonitor monitor, Properties properties) throws DBException {
+        monitor.subTask("Initialising SSL configuration");
+        properties.put(TiDBLakeConstants.SSL_PARAM, "enable");
+    }
+}
